Preallocating space for 'readme' and 'description' is too problematic.
It makes it too hard to find entries that are missing the info, and
too easy to accidentally think that we have more entries with 'readme'
files than we really do.  I'm defaulting new repo entries to use empty
strings for these fields and fixing all the database entries to match.


Former-commit-id: b81b699feb59f0c20f16b9a2c463b80facefd7ae