1) Reduced number of API calls. 2) Added new options. 3) Bugs.
1) I realized that loop() was unnecessarily calling GitHub to find out
the number of API calls left at every turn of the loop, when in fact
it could just keep going until it gets an exception and then check if
the exception was due to hitting the call limit.  This removes a lot
of needless network calls (and is probably nicer to GitHub anyway).

2) Support new option for minimum id number when doing things like
listing entries or indexing readme files.

3) Fixed some bugs found along the way.


Former-commit-id: a2de654bc13f90ec4cedd9350490c064db95275c