1) Get external home page. 2) Get num contributors.
The number of contributors turns out to be difficult to get via HTML,
because the info seems to be updated reactively on GitHub.  I tried
waiting for a while and re-getting the HTML, but it didn't produce
different content.  So for now we'll just have to store it only when
we manage to get it on the first try.