Fixed bug in language getter code.
The value of the 'time' field should be retrieved always, because
update_entry_field(...) sets one of the subfield values.  Although
update_entry_field(...) itself doesn't need the subfields when calling
the mongodb update function, the fact that update_entry_field(...)
then sets the value in the entry structure (for the benefit of the
caller) might cause unexpected problems elsewhere in our code.  Better
to get the data every time and not risk hard-to-track-down bugs in the
future.